update: pokemon-en.jsを作成し、pokemon-ja.jsはpokemon-en.jsを継承するように修正 get-pokemon.jsからpokemonObjへpokemonImgを渡さないように修正
