+++ src/pokemon/get-pokemon.js
@@ -1,6 +1,5 @@
 // @flow
 
-import Pkparaiso from './pkparaiso';
 import PokemonJa from './pokemon-ja';
 import { format } from 'date-fns';
 
@@ -63,7 +62,7 @@ export default class GetPokemon {
         try {
             const pokeSelect = Math.floor(Math.random() * MAX) + 1;
             const data = await this.requestApi.request(pokeSelect);
-            this.pokemonObj = new PokemonJa(data, Pkparaiso);
+            this.pokemonObj = new PokemonJa(data);
             this.time = this.createGetTime();
             return this.createSuccessRes();
         } catch (err) {
